refactor(scorchwrap): extract generator wrapping to a separate file
