+++ packages/webpack/src/ENUM.json
@@ -1,5 +1,7 @@
 {
   "NAME_globalThis": "G",
   "NAME_scopeTerminator": "ST",
-  "NAME_runtimeHandler": "RH"
+  "NAME_runtimeHandler": "RH",
+  "RUNTIME_KEY": "_LM_"
+
 }
